Create custom hook `useObjectUrl`
could be useful for the Add to Apple calendar button
